fix(sim): Make simulation process actually continuous (#86)
* fix(sim): Make simulation process actually continuous

Right now, the simulator stops as soon as there are no more items to simulate. This actually is not correct behavior, as we could be early in the slot, with more items coming in later in the slot which could yield a higher total payout on the block.

What we actually want, is to simply skip doing the round at all if there are no items, sleep for a certain amount of time (bounded by the deadline), and either break (if we slept until the deadline) or continue if there's still time remaining.

The resulting behavior is that the simulator always runs until the deadline, ensuring that items that come later in the slot, up until the deadline, can be simulated and included in the block.

* chore: properly break if we sleep until deadline

* chore: make amount of time in ms that the sim sleeps a constant

* chore: rm dumb comparison

* chore: rm newline

* chore: simplify logic

* chore:
comment

* chore: bias select!