chore(tests): Make timestamp divisible by 12 to conform to slot invariant (#90)
* chore(tests): Make timestamp divisible by 12 to conform to slot invariant

Right now, this timestamp isn't divisible by 12. After fixing the bug we had on blob extraction on the node (https://github.com/init4tech/signet-node/pull/557/files#diff-c630b8329efc76498eaf78bdf3073dcdaefaacb9325e5731b7d1d8cae855d791R262), we now don't conform to an invariant that the calculator has when using `slot_ending_at`: The timestamp must be a slot boundary. We know if it's a slot boundary if it's divisible by 12 in the context of the start timestamp. For tests, this is 0, so any timestamp divisible by 12 will do.

* chore: v bump

* stuff